Move claims list item
